持续更新...

+++ src/main/java/com/fc/controller/MessageController.java
@@ -23,7 +23,7 @@ public class MessageController {
     //去消息页面
     @RequestMapping("/toMessage.do")
     public String toMessage(Model model, HttpSession session) {
-        Integer sessionUid = (Integer) session.getAttribute("uid");
+        Long sessionUid = (Long) session.getAttribute("uid");
         Map<String,List<Message>> map = messageService.listMessageByUid(sessionUid);
         model.addAttribute("map",map);
         System.out.println(map);
